a lot of fixes

+++ Roles/Impostor/Devourer.cs
@@ -116,7 +116,7 @@ namespace EHR.Impostor
             target.SetColor(outfit.ColorId);
             sender.AutoStartRpc(target.NetId, (byte)RpcCalls.SetColor)
                 .Write(target.Data.NetId)
-                .Write(outfit.ColorId)
+                .Write((byte)outfit.ColorId)
                 .EndRpc();
 
             target.SetHat(outfit.HatId, outfit.ColorId);
