+++ Roles/Neutral/Doppelganger.cs
@@ -120,7 +120,7 @@ public class Doppelganger : RoleBase
         pc.SetColor(newOutfit.ColorId);
         sender.AutoStartRpc(pc.NetId, (byte)RpcCalls.SetColor)
             .Write(pc.Data.NetId)
-            .Write(newOutfit.ColorId)
+            .Write((byte)newOutfit.ColorId)
             .EndRpc();
 
         pc.SetHat(newOutfit.HatId, newOutfit.ColorId);
